Moved to util class
